Backend

+++ server/src/api/projects/projects.service.ts
@@ -1,4 +1,5 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TeamsService } from 'src/api/teams/teams.service';
 import { Repository } from 'typeorm';
@@ -12,10 +13,15 @@ export class ProjectsService {
     @InjectRepository(Project)
     private projectRepository: Repository<Project>,
     private teamService: TeamsService,
+    private jwtService: JwtService,
   ) {}
 
-  async create(dto: CreateProjectDto) {
-    const team = await this.teamService.getTeamById(dto.teamId);
+  private async decodeToken(token: string) {
+    return JSON.parse(JSON.stringify(this.jwtService.decode(token)));
+  }
+
+  async create(dto: CreateProjectDto, token: string) {
+    const team = await this.teamService.getTeamById(dto.teamId, token);
 
     if (team) {
       const project = await this.projectRepository.create({
@@ -24,7 +30,7 @@ export class ProjectsService {
       });
       return this.projectRepository.save(project);
     }
-    throw new HttpException('Команда не найдена', HttpStatus.NOT_FOUND);
+    throw new HttpException('Ошибка создания проекта', HttpStatus.NOT_FOUND);
   }
 
   async getAllProjects() {
@@ -35,29 +41,65 @@ export class ProjectsService {
     throw new HttpException('Проекты не найдены', HttpStatus.NOT_FOUND);
   }
 
-  async getProjectById(id: number) {
-    const project = await this.projectRepository.findOne({
-      where: { id },
-      relations: ['tasks'],
-    });
-    if (project) return project;
-    throw new HttpException('Проект не найден', HttpStatus.NOT_FOUND);
-  }
-
-  async delete(id: number): Promise<Project> {
-    const project = await this.projectRepository.findOneBy({ id });
-    if (project) return this.projectRepository.remove(project);
-    throw new HttpException('Проект не найден', HttpStatus.NOT_FOUND);
-  }
-
-  async update(id: number, dto: UpdateProjectDto): Promise<Project> {
-    const project = await this.projectRepository.findOneBy({ id });
-    if (project) {
-      const newProject = await this.projectRepository.merge(project, {
-        name: dto.name,
+  async getProjectById(id: number, token: string) {
+    const decoded = await this.decodeToken(token);
+    if (decoded) {
+      const project = await this.projectRepository.findOne({
+        where: {
+          id,
+          team: {
+            users: {
+              id: decoded.id,
+            },
+          },
+        },
+        relations: ['tasks'],
       });
-      return this.projectRepository.save(newProject);
+      if (project) return project;
+      throw new HttpException('Проект не найден', HttpStatus.NOT_FOUND);
     }
-    throw new HttpException('Проект не найден', HttpStatus.NOT_FOUND);
+    throw new HttpException('Ошибка авторизации', HttpStatus.UNAUTHORIZED);
+  }
+
+  async delete(id: number, token: string): Promise<Project> {
+    const decoded = await this.decodeToken(token);
+    if (decoded) {
+      const project = await this.projectRepository.findOneBy({
+        id,
+        team: {
+          creator: {
+            id: decoded.id,
+          },
+        },
+      });
+      if (project) return this.projectRepository.remove(project);
+      throw new HttpException('Ошибка удаления проекта', HttpStatus.NOT_FOUND);
+    }
+    throw new HttpException('Ошибка авторизации', HttpStatus.UNAUTHORIZED);
+  }
+
+  async update(id: number, dto: UpdateProjectDto, token): Promise<Project> {
+    const decoded = await this.decodeToken(token);
+    if (decoded) {
+      const project = await this.projectRepository.findOneBy({
+        id,
+        team: {
+          creator: {
+            id: decoded.id,
+          },
+        },
+      });
+      if (project) {
+        const newProject = await this.projectRepository.merge(project, {
+          name: dto.name,
+        });
+        return this.projectRepository.save(newProject);
+      }
+      throw new HttpException(
+        'Ошибка обновления проекта',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    throw new HttpException('Ошибка авторизации', HttpStatus.UNAUTHORIZED);
   }
 }
